llama-hf-finetune/requirements.txt
torch>=2.0
transformers>=4.35
datasets>=2.13
scikit-learn
pandas==1.4.2
numpy==1.21.6
pyyaml
docker==5.0.3
jupyter==1.0.0
pytest==7.1.2
wandb==0.12.21
accelerate>=0.22
peft>=0.4.0
bitsandbytes>=0.39.0
evaluate>=0.4
optuna>=3.0
sentencepiece
huggingface_hub
tqdm
rouge-score